# docs/swagger.yaml
basePath: /api/
definitions:
  types.Account:
    properties:
      clerk_id:
        type: string
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      provider:
        type: string
      role:
        type: string
      updated_at:
        type: string
    type: object
...
  types.Admin:
    properties:
      account:
        $ref: '#/definitions/types.Account'
      id:
        type: string
    type: object
...
  types.Customer:
    properties:
      account:
        $ref: '#/definitions/types.Account'
      id:
        type: string
    type: object
  types.DeleteCustomerResponse:
    properties:
      customer:
        $ref: '#/definitions/types.Customer'
      message:
        type: string
      ok:
        type: boolean
    type: object
  types.DeleteEmployeeResponse:
    properties:
      employee:
        $ref: '#/definitions/types.Employee'
      message:
        type: string
      ok:
        type: boolean
    type: object
  types.Employee:
    properties:
      account:
        $ref: '#/definitions/types.Account'
      hire_date:
        type: string
      id:
        type: string
      num_ratings:
        type: integer
      performance_score:
        type: number
      position:
        type: string
      status:
        description: ACTIVE / ONDUTY / INACTIVE
        type: string
    type: object
  types.EmployeeTimesheet:
    properties:
      created_at:
        type: string
      employee_id:
        type: string
      id:
        type: string
      status:
        type: string
      time_in:
        type: string
      time_out:
        type: string
      updated_at:
        type: string
      work_date:
        type: string
    type: object
  types.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  types.GetAllCustomersResponse:
    properties:
      customers:
        items:
          $ref: '#/definitions/types.Customer'
        type: array
    type: object
  types.GetCustomerResponse:
    properties:
      customer:
        $ref: '#/definitions/types.Customer'
    type: object
  types.GetEmployeeResponse:
    properties:
      employee:
        $ref: '#/definitions/types.Employee'
    type: object
...
  types.SignUpAdminRequest:
    properties:
      clerk_id:
        type: string
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      provider:
        type: string
      role:
        type: string
    required:
    - clerk_id
    - email
    - first_name
    - last_name
    - provider
    - role
    type: object
  types.SignUpAdminResponse:
    properties:
      admin:
        $ref: '#/definitions/types.Admin'
    type: object
  types.SignUpCustomerRequest:
    properties:
      clerk_id:
        type: string
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      provider:
        type: string
      role:
        type: string
    required:
    - clerk_id
    - email
    - first_name
    - last_name
    - provider
    - role
    type: object
  types.SignUpCustomerResponse:
    properties:
      customer:
        $ref: '#/definitions/types.Customer'
    type: object
  types.SignUpEmployeeRequest:
    properties:
      clerk_id:
        type: string
      email:
        type: string
      first_name:
        type: string
      hire_date:
        type: string
      last_name:
        type: string
      position:
        type: string
      provider:
        type: string
      role:
        type: string
    required:
    - clerk_id
    - email
    - first_name
    - hire_date
    - last_name
    - position
    - provider
    - role
    type: object
  types.SignUpEmployeeResponse:
    properties:
      employee:
        $ref: '#/definitions/types.Employee'
    type: object
  types.TimeInRequest:
    properties:
      employee_id:
        type: string
      time_in:
        type: string
    type: object
  types.TimeOutRequest:
    properties:
      employee_id:
        type: string
      time_out:
        type: string
    type: object
  types.UpdateCustomerRequest:
    properties:
      customer_id:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
    required:
    - id
    type: object
  types.UpdateCustomerResponse:
    properties:
      customer:
        $ref: '#/definitions/types.Customer'
    type: object
  types.UpdateEmployeeRequest:
    properties:
      email:
        type: string
      employee_id:
        type: string
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
    required:
    - id
    type: object
  types.UpdateEmployeeResponse:
    properties:
      employee:
        $ref: '#/definitions/types.Employee'
    type: object
  types.UpdateEmployeeStatusRequest:
    properties:
      id:
        type: string
      status:
        type: string
    required:
    - id
    - status
    type: object
  types.UpdateEmployeeStatusResponse:
    properties:
      ok:
        type: boolean
    type: object
...
  types.UpdatePerformanceScoreRequest:
    properties:
      id:
        type: string
      score:
        type: number
    required:
    - id
    - score
    type: object
  types.UpdatePerformanceScoreResponse:
    properties:
      ok:
        type: boolean
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is the official API documentation for the Handworks Api.
  title: Handworks API
  version: "1.0"
paths:
  /account/admin/signup:
    post:
      consumes:
      - application/json
      description: Create a new admin account
      parameters:
      - description: Admin signup data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/types.SignUpAdminRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.SignUpAdminResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Sign up a new admin
      tags:
      - Account
  /account/customer:
    get:
      description: Retrieve all customer info
      parameters:
      - default: 0
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.GetAllCustomersResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all customers
      tags:
      - Account
  /account/customer/{id}:
    get:
      description: Retrieve customer info
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.GetCustomerResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a customer by ID
      tags:
      - Account
    put:
      consumes:
      - application/json
      description: Update customer information
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated customer info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/types.UpdateCustomerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.UpdateCustomerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a customer
      tags:
      - Account
  /account/customer/{id}/{accId}:
    delete:
      description: Remove a customer by ID
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: string
      - description: Account ID
        in: path
        name: accId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.DeleteCustomerResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a customer
      tags:
      - Account
  /account/customer/signup:
    post:
      consumes:
      - application/json
      description: Create a new customer account
      parameters:
      - description: Customer signup data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/types.SignUpCustomerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.SignUpCustomerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Sign up a new customer
      tags:
      - Account
  /account/employee:
    get:
      description: Retrieve all customer info
      parameters:
      - default: 0
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.GetAllCustomersResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all customers
      tags:
      - Account
  /account/employee/:
    get:
      description: Retrieve employee info
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.GetEmployeeResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get an employee by ID
      tags:
      - Account
    put:
      consumes:
      - application/json
      description: Update employee information
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated employee data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/types.UpdateEmployeeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.UpdateEmployeeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update an employee
      tags:
      - Account
  /account/employee/{id}/{empId}:
    delete:
      description: Remove employee by ID
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: string
      - description: Account ID
        in: path
        name: accId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.DeleteEmployeeResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete an employee
      tags:
      - Account
  /account/employee/{id}/performance:
    patch:
      consumes:
      - application/json
      description: Adjust performance score
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: string
      - description: New score
        in: body
        name: score
        required: true
        schema:
          $ref: '#/definitions/types.UpdatePerformanceScoreRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.UpdatePerformanceScoreResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update employee performance score
      tags:
      - Account
  /account/employee/{id}/status:
    patch:
      consumes:
      - application/json
      description: Set employee ACTIVE/ONDUTY/INACTIVE
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: string
      - description: New status
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/types.UpdateEmployeeStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.UpdateEmployeeStatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update employee status
      tags:
      - Account
  /account/employee/signup:
    post:
      consumes:
      - application/json
      description: Create a new employee account
      parameters:
      - description: Employee signup data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/types.SignUpEmployeeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.SignUpEmployeeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Sign up a new employee
      tags:
      - Account
  /account/employee/timesheet/timein:
    post:
      consumes:
      - application/json
      description: Record employee time-in for the current day
      parameters:
      - description: Time In Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/types.TimeInRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.EmployeeTimesheet'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Employee time in
      tags:
      - Account
  /account/employee/timesheet/timeout:
    post:
      consumes:
      - application/json
      description: Record employee time-out for the current day
      parameters:
      - description: Time Out Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/types.TimeOutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.EmployeeTimesheet'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Employee time out
      tags:
      - Account
  /account/employee/timesheet/today:
    get:
      description: Retrieve the employee timesheet record for the current day
      parameters:
      - description: Employee ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.EmployeeTimesheet'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get today's timesheet
      tags:
      - Account
...
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer <your_token>"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"